Add extent metadata and write it to disk
New structure, ExtentMeta that holds the metadata that goes with
each extent.  This includes the dirty bit, the flush number, and
two future members (ext_version and gen).

Logic updated to support setting the dirty bit on disk when we
have un-flushed writes.  Flush_number is also incremented when
an extent is flushed to disk.

Updated the extent open implementation to try first to read
existing extent files, and only create them new if they don't
already exist.

Added some comments and fixed a few warnings.