First pass with write and flush commands.
There is a bunch here, but we are not yet at a functional state
for this repo.  The Write, WriteAck, Flush, and FlushAck commands
have been added to the message protocol.  They have some of what
they will need, but not everything.  We don't yet have dependencies
for writes, nor do we keep track of what we have written that the
flush needs.  However, you can issue a write then a flush from
upstairs and the data should end up on the FS below downstairs.

A rudementary work queue (hashmap) has been implemented, and we may
decide to throw it out and do something else.

Also, the logic for translating a write offset to the extent index
is currently in the downstairs, but that needs to move upstairs.

The flush numbers for each extent is not yet working, though the
structure is there.  The dirty bit for upstairs is not implemented
yet.

The upstairs does get the flush numbers from each downstairs, but no
logic is written to compare them, and the first connection provides what
the upstairs holds as the source of truth.

The current upstairs connects to all the downstairs you request then
will step through doing a write then a flush then a write then a
flush.  It's all hard coded at the moment, and that should change soon
as well.

There are a ton of debug messages, and those will be coming out soon
as well.