Improve new work notification to worker tasks.
Update the work allocator and upstairs tasks to use the watch variable as a
doorbell to indicate new work is on the work hashmap.

Update the work structure to include status for each downstairs task
to indicate what state an IO request is in.

Each upstairs worker task will, when it receives the doorbell, walk the
hashmap and look for new work specific for that task/downstairs.  All new work
items are put on a list and that task will then perform that work.

When downstairs has finished and send the finished message back to
the upstairs, the upstairs task then marks that job done.

A future change will include creation of a work completed ringbuf or
something similar and the actual removal of IO tasks from the work
hashmap.

Cleaned up a bunch of debug messages, made a few more debug messages.
Test code still lives in main, which will be addressed soon.