Added layer to (eventually) allow the crucible crate to accept work
There is a new task in crucible with a work queue that will eventually
provide a way for crucible crate users to submit work into crucible.
This task is responsible for taking a generic R/W/F command and
determine which downstairs extent(s) and offset(s) will receive the
command.  The task will translate the command into upstaris specific
IO requests and handle submitting that request to the work queue and
notifying the upstairs tasks about the new request.

Nothing yet is connected to what happens when that request is completed.