Add read/write support for multi-block length operations.
Read and Write operations from upstairs to downstairs now support
multi-block length operations, as long as the offset+length is
still less than the overall extent size.

Updated the read protocol message to include a length for the
read request.

Added a new check function and a bunch of tests to verify that a
block offset plus buffer size will be within the extent.

Fixed an upstairs test that was broken.