New structures methods and functions for Guest to Crucible IO
Crucible now has a Guest structure as part of the Upstairs.  This
structure with methods and new functions supports IO operations
coming from outside with an offset and buffer.  Crucible can now
translate these into the required downstairs information and handle
creating and submitting work to the downstairs connections.

A few other misc comments and some fmt and clippy changes to other
files, but nothing that should change any functionality.