Make upstairs into a Crucible lib.
Moved 98% of the upstairs code into a lib file.
Changed upstairs to have both a bin and a lib.

Created a new client directory with a client binary that makes use
of the new crucible lib (that was formerly upstairs).

Updated the README to reflect the new way to run a sample crucible
workload.

This is the first part of making Crucible ready for being a client
of Propolis.