Move the logic for GuestWork completion (#20)
Made methods on the GuestWork and Work sturctures to directly handle
moving and completing jobs inside Crucible.

Added new methods to Work to return a list of completed jobs.
Added new methods to GtoS to transfer data buffers from read results
from downstairs to the guest provided buffer.

Added mpsc::channel to allow the task that downstairs IO completion to
indicate to a listening upstairs task that all work for a respective
downstairs ID has completed and it can start preparing an ack back to
the Guest.

Refactored all the io specific completions into a more generic method as
part of this, which then let me delete all those functions.

Renamed a few things to better indicate their function or purpose.

Created a specific upstairs task to listen and handle the final stages
of a GuestWork job.  Eventually we will connect this to the notification
that goes back to the guest.