Submit operations to Crucible from a NBD server (#22)
Spin up a nbd server (courtesy of the rust-nbd crate) and send the operations to
crucible.

Implement a notification system - sending an operation to a Guest now returns a
Receiver object, and clients can block on the completion of that operation.

A notable difference here is that BytesMut is not longer used as the data field
in BlockOp::Read - it does not guarantee that memory will be shared between
cloned objects. A new Buffer type is used.

See nbd_server/src/README.md for instructions.