Added dtrace to Crucible
Added dtrace to Crucible with a few simple probes and a sample dtrace
program to show how to use it.

Also added a README.md to the tools directory so we can document what
the various things in that directory do.

NOTE: dtrace requires rustc nightly, so rustup as required.