Keep flush number in upstairs, send to downstairs.
The flush number (the highest from any valid downstairs) is now
stored in the upstairs.  When a flush is issued, this number is
then sent to downstairs to use as the "next" flush number for any
extent that has changed.

This required removing the the flush Vec from the flush command
and instead sending this main flush number in its place.

Updated the Upstairs struct to track this flush number along with
the list of flush numbers received on client connect.  Added a method
to the upstairs struct so a new flush can get a downstairs id and the
next flush number at the same time to avoid any dependency deadlock.

Updated downstairs to respect the flush number received from upstairs
and use that on blocks that have changed and require a flush update.

Fixed a clippy complaint or two.