Move IO functions to methods, and added flush dependencies (#33)
Moved the read/write/flush functions that take an offset/len/buff
that build the upstairs tracking struct and downstairs work struct
into methods on the Upstairs.  This enables better locking hygiene.
It also allows us to collect a dependency list in a place where we
can prevent new jobs from getting in the way.

Added a dependency list to the flush command.  The downstairs currently
ignores this, more on that coming soon.

The "test" crucible-client binary is in transition also, pending
additonal refactor coming soon.