Regions: create or do not, there is no try.
Updated the downstairs code with a new create option for creating
a new region like fileystem with extent files and their initial
contents.

Updated downstairs to expect no files when creating, and to expect
all the proper files when just imporing without creation.

This should avoid accidental downstairs removal, let us see errors
when we would otherwise have ignore them, and pave the way for
an import/export command in the downstairs.