Export all or parts of a downstairs region filesystem
With the export option, we can now take either the entire downstairs
extent files and extract the data into a single file, ignoring any
crucible metadata.

We can also, optionally, select a starting block and/or a count of
blocks to copy if we do not wish to export the entire data set.

Cleaned up a few println's as well.