dtrace probes and fewer versions
Added and updated the dtrace probe points, we now have an entry/exit
point for each IO type: Read, Write, and Flush.  The current points are
when an IO lands on the internal upstairs queue, and when that IO has
been completed and that completion was sent back to the guest.

Cleaned up the flush number information that is printed on startup to only
print the value for the first 12 extents.

Added another dtrace script and updated the tools README.