Update client test program, new test BlockOp.
Refactored and cleaned up the client example test program and made
it a behave a bit better.  The program now takes two different workload
options and will figure out the block size and region size instead
of having it all hard coded.

Deleted a bunch of commented out test code and added command line
options to the program so the desired test can be run.  Refactored the
existing two test functions.

Added a new BlockOp::QueryExtentSize for testing so test programs
can get the extent size which can be useful for testing when boundaries
are crossed.