Dtrace (#43)
* dtrace probes and fewer versions

Added and updated the dtrace probe points, we now have an entry/exit
point for each IO type: Read, Write, and Flush.  The current points are
when an IO lands on the internal upstairs queue, and when that IO has
been completed and that completion was sent back to the guest.

Cleaned up the flush number information that is printed on startup to only
print the value for the first 12 extents.

Added another dtrace script and updated the tools README.

* Update client test program, new test BlockOp.

Refactored and cleaned up the client example test program and made
it a behave a bit better.  The program now takes two different workload
options and will figure out the block size and region size instead
of having it all hard coded.

Deleted a bunch of commented out test code and added command line
options to the program so the desired test can be run.  Refactored the
existing two test functions.

Added a new BlockOp::QueryExtentSize for testing so test programs
can get the extent size which can be useful for testing when boundaries
are crossed.

* Replace 512 with block_size