Dependency support in the downstairs (#50)
The downstairs now honors dependencies sent to it from the upstairs.
All IO commands come with a list of dependencies that must finish before
this IO is allowed to move forward.  The downstairs now has a work
hashmap and all new jobs are put onto this hashmap.  A bit of the
upstairs code was re-used for this when possible.

On recption of a job, the current list of jobs is scanned to see if any
progress can be made on any of them before the new job is attempted.

On completion of a job, the list of outstanding jobs is scanned to see
if any of them can proceed.

Read requests now require a dependency list as well.  We may revisit
this decision in the future, but for now they operate the same way as
the writes and flushes and their dependencies are honored.

A new struct was created to track Upstairs internal communication and a
new flag added to the upstairs for testing that allows it to sometimes
forget to send a job to a downstairs (to force dependency conflicts).

A new testing BlockOp was created to allow a guest to know the current
count of active jobs.

With the new upstairs option, anyone using crucible lib needs to update
their CrucibleOps struct with the new field.