Initial downstairs state work.
The first of probably many change to support keeping a state for each
downstairs in the upstairs.  It is mostly comments at this point with
only some states set.

Added a panic hook to the upstairs-client and the downstairs binaries, so
any async task that panics will panic the entire program.  The upstairs
lib does not set it, as that is left to the user of the library to
decide what they want to do.

Added more comments in the downstairs around the ExtentMeta format.

Updated the names of several channels used for communication between
the upstairs tasks to better reflect their roles.  Updated comments
around the channels to describe what they do.

Broke apart the up_listen() function into two parts, and added a loop
and check in the beginning of up_listen() that waits for all three
downstairs tasks to start, contact their respective downstairs, and
get status back, before up_listen() will start taking work off the
queue from the guest.  Moved the rest of up_listen() into a function
that does the work of Guest->Upstairs IO translation.