UUIDs for downstairs, and new arg parsing
Updated the way the downstairs is started.  You now provide a subcommand
and then options for that subcommand, which results in less global
options that are silently ignored when not used.  This makes both the
starting syntax (and the code I stole, ah, was inspired by) similar to
Omicron sled-agent.

Added a UUID to the downstairs region struct.  This is recorded in the
region.json file for each downstairs.

Added two new messages to the upstairs/downstairs initial communication
that requests the RegionDefinition information, including the new
UUID from each downstairs.  Each downstairs UUID is now recorded in the
Upstairs struct.  More coming here later along with processing the rest
of the region information.

Updated all the READMEs I could find that show how to run the
downstairs.  Updated any test scripts or tools I could find as well.