Errors! (#60)
Introduce CrucibleError type that is returned to the Guest from either
Upstairs or Downstairs.

Keep track of downstairs errors in Work, and mark downstairs as bad if
they return errors. Transition downstairs to Failed and skip sending
them work if they are marked as bad.

Separate IOState and ack status of DownstairsIO.

Bad reads should not cause a Failed transition, but only transfer data
on good read. Bad writes or flushes should cause a failed transition, with
reconciliation and/or retries to follow.