V1 flow control and region matching (#69)
Broke the initial upstairs downstairs negotiation into a different
function.  Moved the loop for new work and responses from downstairs
into its own function.

Added a basic flow control to avoid the current situation where the
upstairs would overwhelm the downstairs with work and not check for
any responses until all the work it had was sent.  This initial flow
control should work until we add credits, or do some more profiling.

Moved the downstairs region passing to the RegionInfo command.  Made
the add_downstairs() method responsible for it and remove it from
both the ExtentVersions command and the process_downstairs()
function.

More polishing of the show_all_work() function.