Upstairs requires activation for multiple connects
In Downstairs, support multiple upstairs connecting at the same time by
spawning off new tokio threads for every connection. During propolis VM
migration we will need to support this.

Upstairs now sends its UUID in the protocol, and the downstairs will use
that to differentiate jobs.

Only one Upstairs can be active at any time. When an Upstairs promotes
itself to active, the downstairs will return errors to the other
previously active Upstairs. Promotion to active is required before
anything else will run.

Compute deadline_secs *inside* tokio::select. Timeouts weren't correct
otherwise.

There is currently no coordination of flush numbers between Upstairs. In
the Upstairs migration scenario I don't think this matters because the
new VM won't be started until the old one is gone, and this will mean
the new Upstairs will read the extent flush numbers when it boots.

The intention is that during a VM migration the target VM will be booted
with the same Upstairs configuration as the source, and will perform the
initial handshake to ensure that it can negotiate and talk to the same
downstairs before allowing the migration to continue.

In Hammer, create 5 CruciblePseudoFiles to test activation handoff. To
the Downstairs, this looks like:

    connection from 127.0.0.1:33764
    connection from 127.0.0.1:33766
    connection from 127.0.0.1:33776
    connection from 127.0.0.1:33778
    connection from 127.0.0.1:33782
    upstairs ce76669d-459a-428e-aab6-77904e792df3 connected
    upstairs 6fbc87a8-64ff-4fb6-b235-17e7f52cb048 connected
    upstairs 3b995d6e-b794-458b-8593-114c9aa063dc connected
    upstairs 518daa3e-4b0f-4077-a723-157a026ae814 connected
    upstairs abbe24d3-e99f-4a8a-9779-9ce8b9085123 connected

    3b995d6e-b794-458b-8593-114c9aa063dc is now active
    Current flush_numbers [0..12]: [0, 0, 0, 0]
    ce76669d-459a-428e-aab6-77904e792df3 is now active
    Current flush_numbers [0..12]: [0, 0, 0, 0]
    518daa3e-4b0f-4077-a723-157a026ae814 is now active
    Current flush_numbers [0..12]: [0, 0, 0, 0]
    abbe24d3-e99f-4a8a-9779-9ce8b9085123 is now active
    Current flush_numbers [0..12]: [0, 0, 0, 0]
    6fbc87a8-64ff-4fb6-b235-17e7f52cb048 is now active
    Current flush_numbers [0..12]: [0, 0, 0, 0]

    upstairs 518daa3e-4b0f-4077-a723-157a026ae814 disconnected
    OK: connection(127.0.0.1:33778): all done
    upstairs abbe24d3-e99f-4a8a-9779-9ce8b9085123 disconnected
    upstairs ce76669d-459a-428e-aab6-77904e792df3 disconnected
    OK: connection(127.0.0.1:33782): all done
    upstairs 6fbc87a8-64ff-4fb6-b235-17e7f52cb048 disconnected
    upstairs 3b995d6e-b794-458b-8593-114c9aa063dc disconnected
    OK: connection(127.0.0.1:33764): all done
    OK: connection(127.0.0.1:33766): all done
    OK: connection(127.0.0.1:33776): all done