Put UUID and ds_state under lock with ds_work
Now the state and UUID for a downstairs all lives behind the same
Mutex as the work related queues.  By doing this, we don't have to
grab two locks when a downstairs changes state, and we can do
all the state change work (failing or updating work for the
downstairs) with that single lock.  If we have a state change, we
had to grab both locks anyway, so this just avoids chances for
a deadlock.  Having a UUID change or downstairs state change should
be a rare situation once everything is online, so it should not
result in any performance impact.

In the struct Upstairs, I renamed ds_work to downstairs, then put
ds_uuid and ds_state inside the new struct.

The struct named Work was also renamed Downstairs as it is no longer
just for work.

The methods for state change and uuid change were already part
of the Upstairs.