Refactor client test, save verify information to file
Added the ability for some (most) of the tests in the client program
to save their internal write counter to a file. Added the ability
to load the internal write counter from a file. This allows us to
run a test in the client program, then record what we think each
block should contain. We can then verify a previous run, and build
on that with the next test.

Made a common function to get region info from the upstairs on
startup, and, if requested, load and verify a write count for
expected data.

Refactored all the client tests to take a region info struct.

Made a few more of the client tests actually verify the data they were
writing out and reading back.

Added a verify in/out test to test_up.sh