Support for a downstairs reconnect to a running upstairs.
This update covers the case where a downstairs that had been connected
and was accepting IO has disconnected, but then reconnected in a short
time window (the actual acceptable disconnect time still TBD).

To support this case, the upstairs will now hold in memory all IOs
between flushes to be able to replay them if a downstairs disconnects
and then reconnects.

The upstairs now tracks the last successful flush id (internal job id,
not the actual flush number) so we can tell it to a downstairs on a
reconnect.

Added crude state machine progression checks.  This should be replaced
with a more formal state machine transition update.  The downstairs
should
get a similar bit of work.

When a downstairs goes away, we walk the active work list and return all
of its jobs back to NEW.  We have to pay special attention to jobs that
are AckReady where the result of this downstairs was part of that
decision
and move the overall job status back to NotAcked.

Added in a flush_needed check.  Now that we keep IOs that are not
covered by a flush in upstairs memory, we also need to check every so
often to see if there are outstanding IOs and submit an internal flush
so we can free that memory.

Brought back the send_work() function that was commented out for so
long.  In the reconnect case, we can fail to send a new work
notification to a downstairs that has just gone away and that condition
is not fatal.

Added a bunch of upstairs work tests to verify the replay stuff works
as expected (for many cases) as well as updated a few tests with new
expected state.

client test program updates: Two new tests, Burst and Generic
Burst: A loop around the demo_workload test.
Send a bunch of work to crucible, wait for it all to be finished (on
all downstairs, not just the ack).
Generic: Do 5000 random R/W/F commands, waiting for the ack for each
command to come back before issuing the next.

Updated the verify test to, if not told to quit, keep running verify in
a loop.

downstairs/src/main:
Added protocol steps for last flush on reconnect path.

protocol:
Added the "send me the last internal flush id" message for telling
a reconnected downstairs the job id of the last flush it did.

tools/downstairs_daemon.sh:
A script for starting all three downstairs then randomly killing one of
them at a time with random sleeps in between killings.