Make Upstairs handoff safer
During promote_to_active, signal to previously active Upstairs that it's
time to shut down. Previously that Upstairs would not be notified that
it's not longer active until it tried to send work. This commit changes
that to be an immediate notification.

As well, the proc thread calling new_work and add_work will need to be
the active Upstairs to grab the Work lock.

Several bug fixes here too:
- draining in-flight work Upstairs didn't work, fixed here
- Upstairs should not need to send a flush if it's not active