Updates to dumping out metadata and comparing data for regions.
Moved the dump command to a new file.

Updated the dump command to print out genration number, flush number
and if the dirty bit set for each extent.

The dump command can now take up to three region directories and will
print out the information for all three.

If the dump command is passed a specific extent to look at, it will
dump the metadata, but also compare up to three regions block by block
and print a summary for each block in an extent which region matches
which other regions.

Updated the downstairs to store the generation number and the version
in the database.  The upstairs does not send them yet, but when it
does we will be ready.

Added some tests for the dump command.