improve image import
This change allows the creation of a region from an image file that is
smaller than the target region size.  If the file is not an exact
multiple of the target block size, the final short block will be
extended with zeroes.

The change also improves the performance of image import by reading a
32MB chunk of the import file at a time, and writing as much of that as
we are allowed into the target extents.

Finally, a small improvement to type safety: using the Block object to
represent the block count we pass to "extent_from_offset()" and adding
some routines for converting to and from byte sizes.