Add more info to ExtentVersions message (#89)
Added the generation number and dirty bit vectors to the
ExtentVersions message. Added two more methods to the region
impl to create and provide this information.

Made a new test for crucible-client that will leave modified extents
without a final flush.
Fixed some new cargo clippy complaints also in crucible-client.

Added the display of generation numbers and dirty bits when a
downstairs connects, but they are not connected to anything yet.

Future commits will make use of generation numbers and the dirty
bits for comparing downstairs.

Added a downstairs dump test to the test_up.sh script.  Not really
anything do to with the rest of this commit.