Make Upstairs handoff safer (#90)
* Make Upstairs handoff safer

During promote_to_active, signal to previously active Upstairs that it's
time to shut down. Previously that Upstairs would not be notified that
it's not longer active until it tried to send work. This commit changes
that to be an immediate notification.

As well, the proc thread calling new_work and add_work will need to be
the active Upstairs to grab the Work lock.

Several bug fixes here too:
- draining in-flight work Upstairs didn't work, fixed here
- Upstairs should not need to send a flush if it's not active

* comment fixes

* move sleep to where it would accurately simulate IO taking longer

* borrow checker wins again

* less indents for work_to_do loop

* restrict to 80 cols

* restrict to 80 cols

* expand comment for work_lock

* panic if downstairs did not clear out old upstairs uuid jobs