DTrace updates for showing Upstairs state
First pass at using DTrace to display some internal Upstairs state, showing
the internal state of the three downstairs as well as a count for the
active jobs of both the guest side and the downstairs side.

Updated comments in the DTrace scripts.

Moved the older DTrace scripts into the newer attribute macro instead
of the build.rs script.  Removed the old build.rs script and d file.