More WIP for ack task.
Building on the previous commit, I've broken out the fw such that
it only lives in the async ack task.  It currently does not actually
send anything back though.

All the work/do_work/etc functions no longer take the framed write
fw.

The more I think about it, the more it seems like a different queue
might be the way to go for acks.  I was trying to re-use the
existing IOop struct, but I'm now reconsidering that, as we also
need to send the result back and that is a field that do_work()
will have that is not currently tracked.