Separate downstairs work task (#102)
Spawn a separate task for doing downstairs work. Before this change the
Downstairs would do work in the same task that reads from the socket,
and this could lead to the Upstairs' socket becoming full if the
Downstairs is processing slower than the Upstairs is sending work. It
also responded to the Upstairs in the same task, which could lead to
further stalls.

This commit uncouples reading, doing work, and responding. There's now a
separate responses queue that is drained by a separate task.

FrameWrite had to change to OwnedWriteHalf and is handed around within
an Arc Mutex. Several places need to write to the socket for other
reasons (like negotiation) and this may change in a future refactor.

This commit also adds job unblocking whenever a job completes, or
whenever a job is added.

This required async-ifying everything because std::sync::Mutex of Work
isn't Send.