merge do_work_task and push_next_jobs
Remove unblock_jobs and push_next_jobs functions and merge their
functionality into do_work_task. do_work_task is now triggered to look
for new work - it no longer has work to do sent to it.

additionally, in resp_loop: pull out of fr.read and push into a queue
right away. have a separate task call proc_frame on the messages. this
unblocks the upstairs that may be waiting on its fw.send.

remove the code from in_progress that returns work if it is in the list
of jobs to do next.

fix up unit tests.