Use Authenticated Encryption
Replace the current AES-XTS-128 with AES-256-GCM-SIV. This builds off
the previous work to add nonces and tags to the protocol.

With authenticated encryption, decryption operations can fail now.
Decrypt in Downstairs::complete and set the job status to a new
DecryptionError if it fails. In the future, maybe we can mark that
Downstairs as untrustworthy? GtoS no longer needs an encryption context.

This means that Downstairs::complete only has to copy Bytes into the
job, not ReadResponses anymore.