Reconciliation and reconnect work
This is a step further along the path to support of reconciliation
during initial upstairs startup.  It's not there yet, but this
does put most of the pieces into a single place where further work
can happen.

Re-did the initial downstairs negotiation to support things coming and
going before the negotiation is complete.  This involved creation of a
ack_requested state that the upstairs holds.  If a downstairs goes
away and comes back before the Upstairs is active, it can indicate
to the downstairs that a activate request was received.  This also
required updating the initial negotiation flow.

Made upstairs keep "looper" task always running.  The only way out is
if there is a panic.

Updated the new tasks in the downstairs to handle disconnects from
upstairs while in the middle of work.

Updated some of the work done to be protected by mutex as more things
are going on at the same time and state can change at any minute.

Moved the "mark all jobs as new" method call to where a downstairs is
reconnected instead of when it was disconnected.  This gives us a
little more time to complete jobs that we have ACKs for.

Added a simple test to run hammer in a loop.