Plumb the generation number all the way to storage
This is part two (of ??) for support of generation number (AKA attach
number) in Crucible.  The generation number supplied during the
attach request for a guest is now passed down through the flush
command all the way to storage.

There is still more to do here, and part of that is design work.