Recorrect buffer's slice range
A previous commit in this PR removed both the beginning and end of
buffer's slice range. this is incorrect but passed all the testing I
did: if Rust's read returns Ok(n), then implementations guarantee that
the number of bytes read is 0 <= n <= buffer.len(). This means that read
can incompletely fill the buffer and the loop must stay in.