Create a simple AWS performance test framework
Bring up an Upstairs plus three Downstairs in an AWS VPC, and run
performance tests against the setup. The basic test so far is simply
running hammer with `--num-upstairs 1` and using encryption.

Both Ubuntu and Helios are supported as the host OS.

The main programs doing the work here are terraform and ansible. Export
your AWS credentials and see README.md for how to run.