Batch downstairs read and write ops (#146)
Benchmark region_read and region_write operations so that many
are sent to a single extent.

Add a criterion benchmark for downstairs. This required adding a
downstairs/src/lib.rs to the project because Criterion cannot use a
binary crate as a dependency.

test_big_write was also added to fully validate region_write and
region_read.