Downstairs reconciliation part 1
Keep track of the region info (gen, flush, dirty) from each downstairs
when they connect.

Added another method to walk the region info and build a list of extents
where the downstairs don't agree with each other.

Removed the process_downstairs function and moved that logic to either
the proc function (to remember the info) or to the mismatch_list()
method where we determine where the issues are.

The default is still to let mismatches continue, as we don't yet have
a way to repair, but that is coming soon.

Added a bunch of tests around finding mismatches with region info.