More reconcile work, create the repair list
Created a new file, mend.rs and moved some of the repair logic into that
file.

Added new downstairs state: Repair.  This will be used when a mismatch
is detected and the upstairs is coordinating the repair to make them
all consistent.

Updated the ds_reconcile method to pave the way for actual repair.
Currently we just build the repair list, but then leave it at that
and continue.

The new mend.rs file contains the logic for both finding a mismatch in
downstairs extents and building the list of repair work that will be
required to make all downstairs consistent.  This repair work consists
of a source extent and one or two destination extents.

Added some checks to allow IO to work if the downstairs is in Repair
state.

Made submit_read and submit_write methods sender field optional.
This opens the door to the upstairs sending IO on its own behalf
instead of only from a guest.  This is similar to what the flush
operation does now.

Simplified the submit_flush and submit_read methods that no longer
need a vec for new work.