Downstairs reconciliation part 1 (#151)
Downstairs reconciliation part 1

Keep track of the region info (gen, flush, dirty) from each downstairs
when they connect.

Added another method to walk the region info and build a list of extents
where the downstairs don't agree with each other.

Removed the process_downstairs function and moved that logic to either
the proc function (to remember the info) or to the mismatch_list()
method where we determine where the issues are.

The default is still to let mismatches continue, as we don't yet have
a way to repair, but that is coming soon.

Added a bunch of tests around finding mismatches with region info.

* More reconcile work, create the repair list

Created a new file, mend.rs and moved some of the repair logic into that
file.

Added new downstairs state: Repair.  This will be used when a mismatch
is detected and the upstairs is coordinating the repair to make them
all consistent.

Updated the ds_reconcile method to pave the way for actual repair.
Currently we just build the repair list, but then leave it at that
and continue.

The new mend.rs file contains the logic for both finding a mismatch in
downstairs extents and building the list of repair work that will be
required to make all downstairs consistent.  This repair work consists
of a source extent and one or two destination extents.

Added some checks to allow IO to work if the downstairs is in Repair
state.

Made submit_read and submit_write methods sender field optional.
This opens the door to the upstairs sending IO on its own behalf
instead of only from a guest.  This is similar to what the flush
operation does now.

Simplified the submit_flush and submit_read methods that no longer
need a vec for new work.