Store multiple encryption contexts (#159)
In order to be crash consistent, we have to store every encryption
context that corresponds to a data block until there's an extent fsync.
This is because sqlite will fsync after every operation but our extents
are only fsynced when the guest submits a flush. This commit sets the
synchronous pragma to FULL to be sure of this.

Add the "counter" column to the encryption context table and increment
it after every new encryption context insertion. Use counter to throw
away everything except the last row when a flush occurs.

Change ReadResponse to contain a list of encryption contexts, an the
upstairs will try each one of them in order to find the one that
decrypts the data successfully.

With multiple encryption contexts, add a detailed block dump