deactivate support part 1 for the Upstairs
New deactivate test for crucible-client
This test runs some IO, then does a deactivate then an activate.
Loops several times.  Verifies data from previous loops.
Added a "count" flag, currently only supported for deactivate test.

New crucible errors for deactivate situations.

lib:
Update what happens when a downstairs tells the upstairs it is no
longer connected.

the process_message task now returns a handle that the main cmd_loop
can select on to detect if that task quits.

Just updated cmd_loop to know when a downstairs task has self
deactavated, and have cmd_loop also exit.

Much better logic in up_listen.  Just one loop now and better handling
of the differences between deactivation and just a single downstairs
going away and coming back.

Added an actual upstairs state instead of just an active boolean.
The supported states are Initializing, Active, and Deactivating.
Having specific states for the upstairs allows deactivate to work and
to tell the difference between initial startup connection, reconnects,
and deactivation.

More logic around activation and deactivation to make sure the
transitions happen at the correct time and report errors when a desired
change is not supported.

Added checks for individual downstairs to determine if they can self
deactivate, when work plus a flush has finished.
Added a check to determine if all downstairs are deactivated and if
so, change the upstairs state to reflect that.

Renamed is_active() to guest_io_ready(). This is used to determine if
we are ready and accepting IO from a guest.

Deactivated state is now used for when downstairs completes the
work required for it to be deactivated.
Disabled is now the state for downstairs when they tell the upstairs
that a new upstairs has connected to them.

New BlockOp: Deactivate
The operation a guest sends to deactivate the upstairs from the
downstairs.

A bunch of tests for deactivation.