Add retry loop for client activate request.
A new option to the client test that will retry the initial client
activate request.  This helps the reconnect test, where it is
possible that downstairs going away and coming back can block the
activate longer than the default internal 10 retires that crucible currently
does.

Updated the test_reconnect.sh script to use the new client option to
retry forever.