Updates for deactivate support
Added deactivate test to test_up.sh script.

Improved the logic around deactivation, get the locks in one
place and short cut deactivation when there is no work in the queue.
Have the deactivation set method also generate a flush to avoid a
race if the command before the deactivation was a flush.