Implement TLS between Upstairs and Downstairs
Add the option to use authenticated TLS between the Upstairs and
Downstairs. This required generalizing the FramedRead and FramedWrite
types so that either TlsStream or TcpStream could be used, plus wrapping
those streams with an enum in order to pass them around without rustc
complaining :)

When I use the term "authenticated" TLS here, I mean that the server
will check that client certificates are signed by a mutual certificate
authority. Clients do **not** yet check that server certs are also
signed by that authority - aka, this is not mutual TLS. I couldn't find
the appropriate options in rustls to enable this.

Scripts that generate keys and certificates have been added to an x509
directory. You'll need the cfssl tool. First, run "make" there to
generate the appropriate files for Upstairs and Downstairs.

Then, create the Downstairs like normal, but run with new arguments:

    cargo run --release -q -p crucible-downstairs -- \
        run -p "44100" -d "disks/d0/" \
        --cert-pem x509/downstairs0.pem \
        --key-pem x509/downstairs0-key.pem \
        --root-cert-pem x509/ca.pem

Certificate number is important here. Part of the TLS connection is a
hostname specification. The Upstairs will try to connect to the first
client as "downstairs0", the second as "downstairs1", and so on.

The Downstairs will then spawn an listener that only accepts client
certificates from the "root" cert. Connect Upstairs with the same
arguments:

    cargo run --bin=crucible-client -- \
        -q \
        -t 127.0.0.1:44100 \
        -t 127.0.0.1:44101 \
        -t 127.0.0.1:44102 \
        --key "HLyo7ZhAf/E9IdX2DDHPHJO0dLgrRxZabWiTlnoKZXc=" \
        --cert-pem x509/upstairs.pem \
        --key-pem x509/upstairs-key.pem \
        --root-cert-pem x509/ca.pem \
        --workload one

To test that certificates not signed by x509/ca.pem are rejected, try
using x509/selfsigned.pem and x509/selfsigned-key.pem in the client
connection. The client should fail with:

        error: InvalidCertificateData("invalid peer certificate: UnknownIssuer")

And the Downstairs should say:

        rejecting connection from 127.0.0.1:49462: Custom { kind: InvalidData, error: AlertReceived(BadCertificate) }

If you try to connect without certificates in the client, it should
fail, and the Downstairs should say:

        rejecting connection from 127.0.0.1:38116: Custom { kind: InvalidData, error: CorruptMessage }